feat: move useReleaseChannel hook into an external file
return the current channel with the set channel method
