+++ ts/components/dialog/debug/ReleaseChannel.tsx
@@ -1,52 +1,9 @@
 import { capitalize } from 'lodash';
-import { useDispatch } from 'react-redux';
-import useUpdate from 'react-use/lib/useUpdate';
-import { localize } from '../../../localization/localeTools';
-import { updateConfirmModal } from '../../../state/ducks/modalDialog';
 import { Flex } from '../../basic/Flex';
-import { SessionButtonColor } from '../../basic/SessionButton';
 import { SessionRadioGroup } from '../../basic/SessionRadioGroup';
 import { HintText } from '../../basic/Text';
-import { ALPHA_CHANNEL, LATEST_CHANNEL, type ReleaseChannels } from '../../../updater/types';
-import { Storage } from '../../../util/storage';
-
-/**
- * Returns a function that can set the release channel to a provided value
- */
-const useReleaseChannel = () => {
-  const dispatch = useDispatch();
-  const forceUpdate = useUpdate();
-
-  return (channel: ReleaseChannels) => {
-    window.log.debug(
-      `[debugMenu] useReleaseChannel Setting release channel to ${channel}. It was ${Storage.get('releaseChannel') || 'not set'}`
-    );
-    dispatch(
-      updateConfirmModal({
-        title: localize('warning').toString(),
-        i18nMessage: { token: 'settingsRestartDescription' },
-        okTheme: SessionButtonColor.Danger,
-        okText: localize('restart').toString(),
-        onClickOk: async () => {
-          try {
-            await Storage.put('releaseChannel', channel);
-          } catch (error) {
-            window.log.warn(
-              `[debugMenu] useReleaseChannel Something went wrong when setting the release channel to ${channel}. It was ${Storage.get('releaseChannel') || 'not set'}:`,
-              error && error.stack ? error.stack : error
-            );
-          } finally {
-            window.restart();
-          }
-        },
-        onClickCancel: () => {
-          dispatch(updateConfirmModal(null));
-          forceUpdate();
-        },
-      })
-    );
-  };
-};
+import { ALPHA_CHANNEL, LATEST_CHANNEL } from '../../../updater/types';
+import { useReleaseChannel } from './hooks/useReleaseChannel';
 
 const items = [
   {
@@ -64,8 +21,7 @@ const items = [
 ];
 
 export const ReleaseChannel = () => {
-  const releaseChannel = Storage.get('releaseChannel') as ReleaseChannels;
-  const setReleaseChannel = useReleaseChannel();
+  const { releaseChannel, setReleaseChannel } = useReleaseChannel();
 
   if (!window.sessionFeatureFlags.useReleaseChannels) {
     return null;
